flask_backend/requirements.txt
Flask
transformers
torch
Pillow
